+++ ext/apm-seata-skywalking-plugin/src/main/java/io/seata/apm/skywalking/plugin/DefaultCoreDoGlobalCommitInterceptor.java
@@ -13,37 +13,26 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.apache.skywalking.apm.plugin.seata;
+package io.seata.apm.skywalking.plugin;
 
-import io.seata.core.protocol.RpcMessage;
-import org.apache.skywalking.apm.agent.core.context.CarrierItem;
-import org.apache.skywalking.apm.agent.core.context.ContextCarrier;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.plugin.seata.common.SWSeataUtils;
+import io.seata.apm.skywalking.plugin.common.SWSeataConstants;
 
 import java.lang.reflect.Method;
 
 /**
  * @author zhaoyuguang
  */
-public class ServerOnRequestProcessorProcessInterceptor implements InstanceMethodsAroundInterceptor {
+public class DefaultCoreDoGlobalCommitInterceptor implements InstanceMethodsAroundInterceptor {
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
                              MethodInterceptResult result) throws Throwable {
-        RpcMessage rpcMessage = (RpcMessage) allArguments[1];
-        String operationName = SWSeataUtils.convertOperationName(rpcMessage);
-        ContextCarrier contextCarrier = new ContextCarrier();
-        CarrierItem next = contextCarrier.items();
-        while (next.hasNext()) {
-            next = next.next();
-            next.setHeadValue(rpcMessage.getHeadMap().get(next.getHeadKey()));
-        }
-        AbstractSpan activeSpan = ContextManager.createEntrySpan(operationName, contextCarrier);
+        AbstractSpan activeSpan = ContextManager.createLocalSpan(SWSeataConstants.SEATA_NAME + "/TC/doGlobalCommit");
 //        activeSpan.setComponent(ComponentsDefine.SEATA);
     }
