+++ ext/apm-seata-skywalking-plugin/src/main/java/io/seata/apm/skywalking/plugin/define/AbstractNettyRemotingInstrumentation.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.apache.skywalking.apm.plugin.seata.define;
+package io.seata.apm.skywalking.plugin.define;
 
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -30,12 +30,11 @@ import static org.apache.skywalking.apm.agent.core.plugin.match.NameMatch.byName
 /**
  * @author zhaoyuguang
  */
+public class AbstractNettyRemotingInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
-public class ClientOnResponseProcessorInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+    private static final String ENHANCE_CLASS = "io.seata.core.rpc.netty.AbstractNettyRemoting";
 
-    private static final String ENHANCE_CLASS = "io.seata.core.rpc.processor.client.ClientOnResponseProcessor";
-
-    private static final String INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.seata.ClientProcessorProcessInterceptor";
+    private static final String INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.seata.NettyRemotingClientSendSyncInterceptor";
 
     @Override
     public ConstructorInterceptPoint[] getConstructorsInterceptPoints() {
@@ -48,7 +47,27 @@ public class ClientOnResponseProcessorInstrumentation extends ClassInstanceMetho
             new InstanceMethodsInterceptPoint() {
                 @Override
                 public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                    return named("process").and(takesArguments(2))
+                    return named("sendSync").and(takesArguments(3))
+                            .and(takesArgument(0, named("io.netty.channel.Channel")))
+                            .and(takesArgument(1, named("io.seata.core.protocol.RpcMessage")))
+                            .and(takesArgument(2, long.class));
+                }
+
+                @Override
+                public String getMethodsInterceptor() {
+                    return INTERCEPTOR_CLASS;
+                }
+
+                @Override
+                public boolean isOverrideArgs() {
+                    return false;
+                }
+            },
+            new InstanceMethodsInterceptPoint() {
+                @Override
+                public ElementMatcher<MethodDescription> getMethodsMatcher() {
+                    return named("sendAsync").and(takesArguments(2))
+                            .and(takesArgument(0, named("io.netty.channel.Channel")))
                             .and(takesArgument(1, named("io.seata.core.protocol.RpcMessage")));
                 }
